parsers with repetition, and with cpslistparsers, @amirsh I think you will like this
