+++ src/main/scala/parsec/Parsers.scala
@@ -1,9 +1,6 @@
 package parsec
 
 import scala.annotation.tailrec
-import scala.collection.mutable.ListBuffer
-
-import barbedwire._
 
 /**
  * An alternative implementation of (unstaged) parser combinators
@@ -68,34 +65,6 @@ trait Parsers {
     def apply[T](f: Input => ParseResult[T]) = new Parser[T] {
       def apply(in: Input) = f(in)
     }
-/*
-    /**
-     * A CPSList from a parser.
-     */
-    def fromParser[T, R](p: Parser[T]) = new CPSList[T, ParseResult[R]] {
-      def fold(z: ParseResult[R], comb: Combine[T, ParseResult[R]]) = {
-
-        @tailrec
-        def loop(curRes: ParseResult[R]): ParseResult[R] = curRes match {
-          //will always be a success here!
-          case s @ Success(_, curIn) => p(curIn) match {
-            /**
-             * This feels a bit convoluted. We only want to combine
-             * the underlying result and propagate the input.
-             * But it makes sense that that is the responsibility
-             * of the `comb` function after all
-             */
-            case s2 @ Success(res, rest) => loop(comb(s, s2))
-            /**
-             * The rest is where we started failing
-             */
-            case Failure(_, _) => s
-          }
-        }
-        loop(z)
-      }
-    }
-    */
   }
 
   /**
@@ -108,9 +77,6 @@ trait Parsers {
   }
 
   def accept(e: Elem) = acceptIf(_ == e)
-
-  def rep[T, R](p: Parser[T])(implicit red: Reducer[T, R]): Parser[R] = Repeat(p)
-
 
   /**
    * An ADT for parsers
@@ -135,47 +101,6 @@ trait Parsers {
   case class Or[T](l: Parser[T], r: Parser[T]) extends Parser[T] {
     def apply(in: Input) = l(in) orElse r(in)
   }
-
-  /**
-   * the repetition parser yields a `R` which is the result type
-   * of a `CPSList`. This helps us compose list operations before
-   * constructing the list.
-   *
-   * We need to have a CPSList hanging around, so we take it as an implicit
-   * @TODO implicits so early seems like an overkill, will need to thinK
-   * this through better.
-   *
-   * In fact we can't have CPSList. `rep` is not a Parser[CPSList], nor is it
-   * a CPSList[Parser]. In fact it is both at the same time! IE a "transformed"
-   * monad of both.
-   * Either we make CPSList monadic (cumbersome), or we go back to
-   * a proper `def repFold[T, R](p: Parser[T])(z: R, acc: (R, T) => R): Parser[R]`
-   *
-   * Let's make the (z, acc) parameter list implicit (so we don't need to specify),
-   * give it a name (`Reducer`) and make an API for it (so we can map, filter etc. over it.)
-   * Attention, `flatMap` will most prolly not work.
-   *
-   * Design of `Reducer` inspired by fastparse, cf. `Repeater`
-   * https://github.com/lihaoyi/fastparse/blob/master/fastparse/shared/src/main/scala/fastparse/ParserApi.scala
-   */
-  case class Repeat[T, R](p: Parser[T])(implicit red: Reducer[T, R]) extends Parser[R] {
-    def apply(in: Input) = {
-      @tailrec
-      def loop(curIn: Input, curRes: R): ParseResult[R] =
-        p(curIn) match {
-          case Success(res, rest) => loop(rest, red.combine(curRes, res))
-
-          /**
-           * The rest is where we started failing
-           */
-          case Failure(_, _) => Success(curRes, curIn)
-        }
-
-      loop(in, red.z)
-    }
-  }
-
-
 
   /**
    * ParseResult, follows the implementation in the standard library
@@ -204,8 +129,3 @@ trait Parsers {
   case class Failure(errMessage: String, rest: Input) extends ParseResult[Nothing]
 
 }
-
-/**
- * A simple Reducer for now
- */
- case class Reducer[A, R](z: R, combine: (R, A) => R)
